feat: add ComponentList component, ensure correct status behavior
